Function to unify coordinates of multiple configurations

Combines multiple configurations from dimension reduction solutions applied to multiple imputed data sets

Usage

GPAbin(missbp, G.target = NULL)

Arguments

missbp

An object of class missmi obtained from preceding function missmi()

G.target

Target configuration. If not specified the centroid configuration will be used as the target.

Value

Z.GPA.list

List containing the sample coordinates for each MI after GPA

CLP.GPA.list

List containing the CLPs for each MI after GPA

G.target

Target configuration

Z.GPAbin

Sample coordinates for the GPAbin biplot

CLP.GPAbin

CLPs for the GPAbin biplot

Examples

data(implist)
missbp <- missmi(implist) |> DRT() |> GPAbin()