Multiple imputation

Choose between four available multiple imputation strategies in R.

Usage

impute(missbp, imp.method = c("MIMCA", "jomo", "DPMPM", "mice"), m = 5)

Arguments

missbp

An object of class missmi obtained from preceding function missmi().

imp.method

Select one of four imputation methods: MIMCA, jomo, DPMPM, mice

m

Number of multiple imputations

Value

The missbp object is appended with the following object:

dataimp

List of imputed data

See also MIMCA, jomo1cat, mi and mice.

Examples

# \donttest{
data(missdat)
missbp <- missmi(missdat) |> impute(imp.method="DPMPM", m=5)# }